refactor($compile): 상품가격 정적 팩토리 메서드 추가(Long형 인자)

+++ src/test/java/kitchenpos/products/tobe/domain/PriceTest.java
@@ -15,7 +15,7 @@ class PriceTest {
   @ValueSource(longs = {0, 16_000L})
   void createPrice(long value) {
     Price price = Price.from(BigDecimal.valueOf(value));
-    assertThat(price).isEqualTo(Price.from(BigDecimal.valueOf(value)));
+    assertThat(price).isEqualTo(Price.from(value));
   }
 
 }
